Search Tab: Dropdown for Department
- replaced department input field with department drop down

**BUGS
- when you select a department than search for a first/last name that’s
not in the department it will show no results (as it should) but not
the ‘no results’ section. Pagination still picks up if there’s a last
name in the JSON file, even though there’s not for the department
filter.
- the order by buttons do not work properly when you only select a
department.

Personally think that the User Experience would be better with a
department dropdown over the department input field - thoughts?